Refactor: Implement Post APIs with React Query
